auth_app/recognizor/requirements.txt
PyArabic
arabic-reshaper
einops
fire
gast
h5py
imageio
kiwisolver
lmdb
Markdown
matplotlib
natsort
nbclient
networkx
nltk
numpy
opencv-python
pillow
scikit-image
scipy
six
timm
torch
torchvision
tqdm
urllib3
validators
Wand
opencv-python==4.5.1.48
opencv-contrib-python==4.5.1.48